Added a tutorial which focuses on multi-state classification on the cardiotocography dataset. (#381)
* Update README.md

Added some benefits of using PyCaret library.

* Update README.md

* Add files via upload

This tutorial gives an overview of how to use pycaret library for multi-class classification, compare and build different classification models, tuning the hyperparameters of a model, saving the model for future use, and loading the saved model. The dataset used in this tutorial is a cardiotcography dataset. We have tried to predict the fetal state using the cardiotocography data. Cardiotocography is the technique that helps doctors to trace the heart rate of the fetus, which includes measuring accelerations, decelerations, and variability, with the help of uterine contractions. Further, this cardiotocography can be used to classify a fetus into three states namely:
Normal trace, Suspicious trace, and Pathological trace.

Co-authored-by: PyCaret <moez@pycaret.org>