Updated the index.csv (#386)
* Update README.md

Added some benefits of using PyCaret library.

* Update README.md

* Add files via upload

This tutorial gives an overview of how to use pycaret library for multi-class classification, compare and build different classification models, tuning the hyperparameters of a model, saving the model for future use, and loading the saved model. The dataset used in this tutorial is a cardiotcography dataset. We have tried to predict the fetal state using the cardiotocography data. Cardiotocography is the technique that helps doctors to trace the heart rate of the fetus, which includes measuring accelerations, decelerations, and variability, with the help of uterine contractions. Further, this cardiotocography can be used to classify a fetus into three states namely:
Normal trace, Suspicious trace, and Pathological trace.

* Add files via upload

2126 fetal cardiotocograms (CTGs) were automatically processed and the respective diagnostic features measured. The CTGs were also classified by three expert obstetricians and a consensus classification label assigned to each of them. Classification was both with respect to a morphologic pattern (A, B, C. …) and to a fetal state (N, S, P). Therefore the dataset can be used either for 10-class or 3-class experiments. This dataset has been preprocessed by me and has been converted into CSV(comma-separated values format). 
Link to the dataset and its details: https://archive.ics.uci.edu/ml/datasets/Cardiotocography

* Updated index.csv

Co-authored-by: PyCaret <moez@pycaret.org>