Fix dependency: Joblib 1.4.0 breaks PyCaret
PyCaret uses the private function `joblib.memory._format_load_msg` from
Joblib, which is no longer present in Joblib 1.4.

In order to fix the problem without much efforts, this patch vendorizes
the corresponding function from Joblib without much ado.

It is advisable to release this update on behalf of a bugfix release, so
that downstream users will get this issue resolved quickly.