Fix dependency: Joblib 1.4.0 breaks PyCaret
PyCaret uses private functions from Joblib, which are no longer present
in Joblib 1.4.

- joblib.memory._format_load_msg
- FastMemorizedFunc._get_output_identifiers

In order to fix the problem without much efforts, this patch downgrades
to joblib<1.4 to remedy the problem as fast as possible.

It is advisable to release this update on behalf of a bugfix release, so
that downstream users will get this issue resolved quickly.